requirements.txt
python-binance
pandas
python-dotenv
